fix photo viewer

+++ app/src/main/java/com/snt/phoney/domain/model/Career.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 
 @Serializable
-data class Career(var id: Int = 0,
+data class Career(var id: Long = 0,
                   var name: String? = null) {
 
     @Transient
