+++ app/src/main/java/com/snt/phoney/domain/model/ReportReason.kt
@@ -3,7 +3,7 @@ package com.snt.phoney.domain.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ReportReason(var id: Int = 0,
+data class ReportReason(var id: Long = 0,
                         var name: String? = null)
